requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anthropic==0.42.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.7.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.24.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.24.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.156.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.37.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-texttospeech==2.23.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-translate==3.19.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.66.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.68.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.68.1 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.4 ; python_version >= "3.11" and python_version < "4.0"
pydub==0.25.1 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
pypdf2==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
